feat: Add cube links to Statistics page
