+++ CHANGELOG.md
@@ -11,7 +11,11 @@ You can also check the
 
 # Unreleased
 
-- Extended maximum URL length in Node to 64KB
+- Features
+  - Valid cube iris in Statistics page (coming from the same data source as the
+    environment's default one) are now pointing to Visualize dataset previews
+- Maintenance
+  - Extended maximum URL length in Node to 64KB
 
 # [5.4.1] - 2025-03-18
